sixth commit

+++ src/main/java/org/example/service/JugadorService.java
@@ -1,24 +1,30 @@
 package org.example.service;
 
+import org.example.controllers.dto.JugadorDto;
+import org.example.dao.EquipoDAO;
 import org.example.dao.JugadorDAO;
+import org.example.entity.Equipo;
 import org.example.entity.Jugador;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.stream.Collectors;
+
 @Service
 public class JugadorService {
-
     @Autowired
     private JugadorDAO jugadorDAO;
+    @Autowired
+    private EquipoDAO equipoDAO;
 
-    public List<Jugador> listJugador() {
-        return jugadorDAO.findAll();
+    public List<JugadorDto> listJugador() {
+        List<Jugador> jugadores = jugadorDAO.findAll();
+        return jugadores.stream().map(JugadorDto::toDto).collect(Collectors.toList());
     }
-
-    public boolean addJugador(Jugador entity) {
+    public boolean addJugador(JugadorDto jugadorDto) {
         try {
-            jugadorDAO.save(entity);
+            jugadorDAO.save(jugadorDto.toEntity(jugadorDto));
             return true;
         } catch (Exception e) {
             return false;
@@ -29,16 +35,33 @@ public class JugadorService {
         return jugadorDAO.findById(jugadorId).orElse(null);
     }
 
+    public void updateJugador(Jugador jugador, JugadorDto jugadorDto) {
+        jugador.setId(jugadorDto.getId());
+        jugador.setNombre(jugadorDto.getNombre());
+        jugador.setDorsal(jugadorDto.getDorsal());
+        jugadorDAO.save(jugador);
+    }
+
     public boolean deleteJugador(int jugadorId) {
         try {
-            jugadorDAO.deleteById(jugadorId);
+            Jugador jugador = findJugador(jugadorId);
+            if (jugador != null) {
+                Equipo equipo = jugador.getIdEquipo();
+                List<Jugador> jugadoresEnEquipo = listJugadoresPorEquipo(equipo.getId());
+                if (jugadoresEnEquipo.size() == 1) {
+                    jugadorDAO.deleteById(jugadorId);
+                    equipoDAO.deleteById(equipo.getId());
+                } else {
+                    jugadorDAO.deleteById(jugadorId);
+                }
+            }
             return true;
         } catch (Exception e) {
             return false;
         }
     }
-
     public List<Jugador> listJugadoresPorEquipo(int equipoId) {
         return jugadorDAO.findByIdEquipo_Id(equipoId);
     }
+
 }
